refactor: move overlay view models into presentation
