+++ extension/src/presentation/content/overlay/view-state.ts
